+++ tests/test_helpers/task_logger.py
@@ -3,7 +3,6 @@ from typing import Any, cast
 import anyio
 
 from inspect_ai._eval.task.log import TaskLogger
-from inspect_ai.dataset._util import SampleKeyLookup
 from inspect_ai.log._recorders.buffer.database import SampleBufferDatabase
 
 
@@ -16,7 +15,5 @@ class TaskLoggerShim(TaskLogger):
         self._init_stale_flush_state()
         self._finished = False
         self._prior_seeded = False
-        self._prior = None
         self._prior_read_limit = anyio.Semaphore(4)
-        self._seeded_lookup = SampleKeyLookup()
         self._seeded_pending = set()